requirements.txt
streamlit
torch
numpy
joblib
ultralytics
pillow
xgboost
